web: Extract document.embeds polyfill to its own file
